requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.1.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
